Sorry i forgot to commit as i went on :P
I added the functionality to delete a project in the repository and added the correct mapping in the controller.
I added tests to see if it correctly deletes the project from the database and to see if, once deleted, it correctly redirects to /projects page again.

In the projects.html i changed it from an <a> tag to a <form> tag as this is more secure.